+++ compatibility-api/src/main/java/nbbrd/compatibility/Filter.java
@@ -1,7 +1,7 @@
 package nbbrd.compatibility;
 
 import lombok.NonNull;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 
 import java.time.LocalDate;
 import java.time.Year;