+++ compatibility-api/src/main/java/nbbrd/compatibility/ReportItem.java
@@ -1,7 +1,7 @@
 package nbbrd.compatibility;
 
 import lombok.NonNull;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 
 import java.net.URI;