+++ compatibility-api/src/main/java/nbbrd/compatibility/spi/Maven.java
@@ -3,7 +3,7 @@ package nbbrd.compatibility.spi;
 import lombok.NonNull;
 import nbbrd.compatibility.Artifact;
 import nbbrd.compatibility.Version;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -22,4 +22,7 @@ public interface Maven {
     Version getArtifactVersion(@NonNull Path project, @NonNull Artifact artifact) throws IOException;
 
     void setArtifactVersion(@NonNull Path project, @NonNull Artifact artifact, @NonNull Version version) throws IOException;
+
+    @Nullable
+    Version getArtifactLatestRelease(@NonNull Artifact artifact) throws IOException;
 }