+++ compatibility-maven-plugin/src/main/java/nbbrd/compatibility/maven/plugin/AbstractCheckStreamMojo.java
@@ -5,7 +5,7 @@ import lombok.NonNull;
 import nbbrd.compatibility.*;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.Parameter;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 
 import java.io.File;
 import java.io.IOException;